refactor: move to existing RestApi namespace
