+++ MailerQ/RestApi/Client/IApiClient.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MailerQ.ApiClient
+namespace MailerQ.RestApi.Client
 {
     public interface IApiClient
     {
