+++ MailerQ/RestApi/Error.cs
@@ -12,6 +12,11 @@ namespace MailerQ.RestApi
     public class Error
     {
         /// <summary>
+        /// Endpoint name
+        /// </summary>
+        public readonly string Endpoint = "errors";
+
+        /// <summary>
         /// Numeric error code between 200 and 599 (smtp error codes).
         /// </summary>
         [JsonProperty(Required = Required.Always)]
