+++ MailerQ/RestApi/Pause.cs
@@ -12,6 +12,11 @@ namespace MailerQ.RestApi
     public class Pause
     {
         /// <summary>
+        /// Endpoint name
+        /// </summary>
+        public readonly string Endpoint = "pauses";
+
+        /// <summary>
         /// Pool that the pause applies to.
         /// </summary>
         public string Pool { get; set; }
